requirements.txt
aws-cdk.core==1.131.0
aws-cdk.assertions==1.131.0
aws-cdk.aws_ecs==1.131.0
aws-cdk.aws_ec2==1.131.0
aws-cdk.aws_ecs_patterns==1.131.0
